requirements.txt
# ── Core CV / ML ──────────────────────────────────────────────────────────────
ultralytics>=8.0.0          # YOLOv8 training + ONNX export (Colab only)
torch>=2.0.0                # PyTorch (Colab only for training)
torchvision>=0.15.0         # torchvision (Colab only)
onnxruntime-gpu>=1.16.0     # ONNX inference with CUDA (local inference)
opencv-python>=4.8.0        # Frame capture + image ops
supervision>=0.16.0         # ByteTrack, annotations, zone/line counting
numpy>=1.24.0

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
pydantic>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6     # FastAPI file uploads
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
Pillow>=10.0.0
scipy>=1.11.0               # Rolling stats for anomaly detection

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0               # async test client for FastAPI
